fall_2025/CV/pyproject.toml
[project]
name = "gesture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastparquet>=2024.11.0",
    "jupyter>=1.1.1",
    "marimo>=0.16.5",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "torch>=2.8.0",
    "wandb>=0.22.3",
]
